[proliferate] split workspace UI store actions
