+++ desktop/src/stores/preferences/workspace-ui-state-value.ts
@@ -1,0 +1,7 @@
+import type { SetStateAction } from "react";
+
+export function resolveStateValue<T>(value: SetStateAction<T>, current: T): T {
+  return typeof value === "function"
+    ? (value as (previousValue: T) => T)(current)
+    : value;
+}
